Move groupby tests to the right module
